Remove try-except ladder
We were using a try-except "ladder" to handle switching between
backends. While there's nothing totally wrong with this, it was allowing for a
weird issue to develop. I had previously installed PyQt5 via my distro package
(Fedora) but had then removed this. Unfortunately removing the package had left
some files behind:

  $ ls /usr/lib64/python3.10/site-packages/PyQt5/* -la
  /usr/lib64/python3.10/site-packages/PyQt5/bindings:
  total 8
  drwxr-xr-x. 2 root root 4096 May 18 15:09 .
  drwxr-xr-x. 4 root root 4096 May 18 15:09 ..

  /usr/lib64/python3.10/site-packages/PyQt5/__pycache__:
  total 8
  drwxr-xr-x. 2 root root 4096 May 18 15:09 .
  drwxr-xr-x. 4 root root 4096 May 18 15:09 ..

This meant that I was able to import the 'PyQt5' module but nothing under this:

  $ python
  >>> import PyQt5
  >>> from PyQt5.QtWidgets import QApplication
  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  ModuleNotFoundError: No module named 'PyQt5.QtWidgets'

This isn't pyperclip's fault. However, by moving away from this
try-catch ladder and wrapping the import logic inside the try block, we
can mitigate against this issue when it happens *and* result in less
indented code.

Signed-off-by: Stephen Finucane <stephen@that.guru>